cart functionalities updated

+++ cart.py
@@ -63,7 +63,7 @@ class Cart():
              
              
     def get_total(self):
-        return sum(Decimal(item['price']) * item['qty'] for item in self.cart.values())
+        return int(sum(Decimal(item['price']) * item['qty'] for item in self.cart.values()))
     
     def delete(self, product):
